Attemted to fix printing issue
Hi you all,
I spent more than five hours trying to figure out how to stop printing the word by replacing it but I could not figure out the issue. I know part of the issue is on the while loop in line173-183. We also have the issue the loop is not adding the past words into version of the word. It is only printing the way the loop is passing through the word. In line 204 to 214 there is another issue, the loop is adding every attempt not just the failed attempts. I attempted to fixed it but I honestly made a non-functioning code. Something that I noticed that we are not keeping track of "word", "chosen-word", "word-guessed". It seems that we are using them interchangebly. I went through the loops and tried to figure out.  The code that I am uploding is not progress code, its frusted code, the issue continued to persist. I am unsure if we need to use .join. Looking forward talking to you all tomorrow.